+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add opt-in Homepage Docker-label fallbacks with `HOMEPAGE_LABEL_FALLBACK=true`.
+
 ## [0.6.14] - 2026-09-06
 
 ### Fixed